msm-poweroff: Remove unused variables
In 9d77fb1 ("msm-poweroff: always preserve PMIC reset status"), we
eliminated the primary function of the need_warm_reset and
get_dload_mode variables - to detect whether or not download mode is
enabled and determine if a warm reset should be performed. Because we
are now forcing a warm reset regardless of CONFIG_QCOM_DLOAD_MODE
status, the necessity of this check has been nullified.

This silences the following compiler warnings:

../drivers/power/reset/msm-poweroff.c:285:7: warning: unused variable
      'need_warm_reset' [-Wunused-variable]
        bool need_warm_reset = false;
             ^
../drivers/power/reset/msm-poweroff.c:152:13: warning: unused function
      'get_dload_mode' [-Wunused-function]
static bool get_dload_mode(void)
            ^

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: mcdachpappe <noreference@web.de>