qpnp-fg-gen3: Fix fg_sram_write() call in clear_cycle_counter()
This not only passes the incorrect size to fg_sram_write(), but also
uses incorrect syntax to pass the pointer of the payload to be written.
`chip->cyc_ctr.count` is defined as an array so there shouldn't be an
ampersand in front of it to get its address. We can also get its size
with just a plain sizeof().

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>