treewide: Remove/clean up even more unnecessary OnePlus changes
In favor of our attempts in the past:
* [1/2] devfreq/drm: Kill off OnePlus's QoS framework: commit b2ed1d51d81a964266754d3712e69b01baac5156
* [2/2] PM / sleep: Kill off OnePlus's QoS framework: commit 7b086d8d755eefec808f475e5a925e7c3cb1968f
* treewide: Remove unused OnePlus scheduling modifications: commit 98d0e7eebffddc5446e1d358ea73b40b980dcd85
* treewide: remove ZRAM, SMART_BOOST, MEMPLUS, SWAP: commit 9af4a930c02d6436eb818c57319135a476b8ce9f

Let's finish this approach of removing all the OnePlus related codelines
and drivers which we don't need and don't want. Please note that hardware/
device detection isn't working anymore - McLaren edition won't be detected
anymore.

While we are at this: Move input/fingerint drivers where they belong to.

Signed-off-by: mcdachpappe <noreference@web.de>